Fix 'content is not allowed in prolog'
Switch order of probing the virtualizerconfigs to bypass xml parsing of
docker configs which are tgz-archives.

We have no good handling of the different config types. VBox and Qemu
are xml, vmware is plaintext, docker wraps everything in tgz. Instead
of checking we actually just try and hope for the best :(.

fixes: #3932